src/requirements.txt
sqlalchemy==1.3.19
psycopg2-binary==2.8.6
nltk==3.5
numpy==1.19.4
pandas==1.1.4
redis==3.5.3
regex==2020.11.13
requests==2.22.0
scipy==1.5.4
sklearn==0.0
stop-words==2018.7.23
pymystem3==0.2.0